+++ src/https/controllers/events/create-event-controller.ts
@@ -1,0 +1,38 @@
+import { FastifyReply, FastifyRequest } from 'fastify'
+
+import { prisma } from '@/lib/prisma'
+import { generateSlug } from '@/utils/generete-slug'
+
+import type { CreateEventData } from './create-event-schema'
+
+export async function CreateEvent(
+  request: FastifyRequest<{ Body: CreateEventData }>,
+  reply: FastifyReply,
+) {
+  const { details, maximumAttendees, title } = request.body
+
+  const slug = generateSlug(title)
+
+  const eventWithSaneSlug = await prisma.event.findUnique({
+    where: {
+      slug,
+    },
+  })
+
+  if (eventWithSaneSlug !== null) {
+    throw new Error('Another event with same title already exists.')
+  }
+
+  const event = await prisma.event.create({
+    data: {
+      title,
+      details,
+      maximumAttendees,
+      slug,
+    },
+  })
+  return reply.status(201).send({
+    message: 'dados cadastrados com sucesso',
+    eventId: event.id,
+  })
+}
