+++ src/lib/prisma.ts
@@ -1,0 +1,5 @@
+import { PrismaClient } from '@prisma/client'
+
+export const prisma = new PrismaClient({
+  log: ['query'],
+})
